refactor: BaseService use new status enum
